+++ winter/web/configurer.py
@@ -1,6 +1,6 @@
 import abc
 
-from .controller import get_instance
+from winter.core import get_injector
 from .interceptor import InterceptorRegistry
 from .interceptor import interceptor_registry
 
@@ -12,6 +12,7 @@ class Configurer(abc.ABC):
 
 
 def run_configurers():
+    injector = get_injector()
     for configurer_class in Configurer.__subclasses__():
-        configurer = get_instance(configurer_class)
+        configurer = injector.get(configurer_class)
         configurer.add_interceptors(interceptor_registry)