+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.0.0] - 2021-03-31
+
+### New features
+- CRUDRepository now supports custom implementation extensions
+### Changed
+- set_factory changed to set_injector
+
 ## [8.2.0] - 2021-01-27
 
 ### New features